ISPN-14749 Combine BackupManagerIT and CliBackupMaangerIT
